pyproject.toml
[project]
name = "bitirme"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "deepagents>=0.4.12",
    "langchain-google-genai>=4.2.1",
    "langgraph-cli[inmem]>=0.4.19",
    "python-dotenv>=1.2.2",
]
